backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.2.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0